+++ src/Routes/indexRoutes.js
@@ -1,0 +1,9 @@
+import { Router } from "express";
+import authRouter from "./authRoutes.js";
+
+
+const router = Router();
+router.use(authRouter);
+
+
+export default router;
